feat : 일정 현재 참가 인원 반환 및 초과 시 예외 처리
