+++ src/main/java/com/kakaotechcampus/team16be/group/repository/GroupRepository.java
@@ -2,6 +2,7 @@ package com.kakaotechcampus.team16be.group.repository;
 
 import com.kakaotechcampus.team16be.group.domain.Group;
 import com.kakaotechcampus.team16be.user.domain.User;
+import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
