+++ src/main/java/com/kakaotechcampus/team16be/plan/service/PlanServiceImpl.java
@@ -12,11 +12,12 @@ import com.kakaotechcampus.team16be.plan.dto.PlanRequestDto;
 import com.kakaotechcampus.team16be.plan.dto.PlanResponseDto;
 import com.kakaotechcampus.team16be.plan.exception.PlanErrorCode;
 import com.kakaotechcampus.team16be.plan.exception.PlanException;
+import com.kakaotechcampus.team16be.planParticipant.dto.PlanParticipantResponseDto;
+import com.kakaotechcampus.team16be.planParticipant.service.PlanParticipantService;
 import com.kakaotechcampus.team16be.user.domain.User;
 import java.time.LocalDateTime;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
-import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -30,29 +31,7 @@ public class PlanServiceImpl implements PlanService {
     private final GroupService groupService;
     private final NotificationService notificationService;
     private final GroupMemberService groupMemberService;
-    private final S3UploadPresignedUrlService s3UploadPresignedUrlService;
-    private final ApplicationEventPublisher eventPublisher;
 
-    @Override
-    public PlanResponseDto getPlan(Long groupId, Long planId) {
-        Plan plan = planRepository.findByGroupIdAndId(groupId, planId)
-                .orElseThrow(() -> new PlanException(PlanErrorCode.PLAN_NOT_FOUND));
-
-        String fullUrl = s3UploadPresignedUrlService.getPublicUrl(plan.getCoverImg());
-
-        return PlanResponseDto.from(plan, fullUrl);
-    }
-
-    @Override
-    public List<PlanResponseDto> getAllPlans(Long groupId) {
-        return planRepository.findByGroupId(groupId)
-                .stream()
-                .map(plan -> {
-                    String fullUrl = s3UploadPresignedUrlService.getPublicUrl(plan.getCoverImg());
-                    return PlanResponseDto.from(plan, fullUrl);
-                })
-                .toList();
-    }
 
     @Override
     @Transactional
